# app/resources/swagger/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LOTR API",
        "description": "Description of LOTR API",
        "contact": {
            "email": "admin@admin.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "url",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/v1/factions": {
            "get": {
                "tags": [
                    "Factions"
                ],
                "summary": "Retrieve a paginated list of factions",
                "operationId": "fe977d45abb091a5d90329fc87a6b077",
                "responses": {
                    "200": {
                        "description": "Successful retrieval of factions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FactionResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unable to retrieve factions"
                    }
                }
            },
            "post": {
                "tags": [
                    "Factions"
                ],
                "summary": "Store a newly created faction",
                "operationId": "81b16f38b90aa8fa6c3fa78158588bb7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FactionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Faction created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FactionResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unable to create faction"
                    }
                }
            }
        },
        "/api/v1/factions/{id}": {
            "get": {
                "tags": [
                    "Factions"
                ],
                "summary": "Retrieve a specific faction by its ID",
                "operationId": "87f4d2c56a36f6ebab9e21a8e3f3fcbb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the faction to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful retrieval of faction",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FactionResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Faction not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Factions"
                ],
                "summary": "Update the specified faction",
                "operationId": "d912866c4a8a0c3c49ea0eec134d4549",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the faction to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FactionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Faction updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FactionResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unable to update faction"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Factions"
                ],
                "summary": "Soft delete the specified faction",
                "operationId": "f14ef7e48478d564062449f39151b426",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the faction to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Faction deleted successfully"
                    },
                    "400": {
                        "description": "Unable to delete faction"
                    }
                }
            }
        },
        "/api/v1/factions/{id}/restore": {
            "post": {
                "tags": [
                    "Factions"
                ],
                "summary": "Restore a soft-deleted faction",
                "operationId": "455454a61cd411a8e78791b33205b003",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the faction to restore",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Faction restored successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FactionResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Faction not found"
                    }
                }
            }
        },
        "/api/v1/factions/{id}/force": {
            "delete": {
                "tags": [
                    "Factions"
                ],
                "summary": "Permanently delete a faction",
                "operationId": "10e93fe86abfb65f78b870582510df27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the faction to permanently delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Faction permanently deleted"
                    },
                    "404": {
                        "description": "Faction not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "FactionRequest": {
                "required": [
                    "faction_name",
                    "description"
                ],
                "properties": {
                    "faction_name": {
                        "type": "string",
                        "maxLength": 128,
                        "example": "The Fellowship of the Ring"
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 65535,
                        "example": "A group formed to destroy the One Ring."
                    }
                },
                "type": "object"
            },
...
            "FactionResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "faction_name": {
                        "type": "string",
                        "example": "The Fellowship of the Ring"
                    },
                    "description": {
                        "type": "string",
                        "example": "A group formed to destroy the One Ring."
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-02T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}